+++ CSharpCodeAnalyst/InfoPanel/InfoPanel.xaml
@@ -1,12 +1,12 @@
 <UserControl x:Class="CSharpCodeAnalyst.InfoPanel.InfoPanel"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:converters="clr-namespace:CSharpCodeAnalyst.Converters"
-             xmlns:cSharpCodeAnalyst="clr-namespace:CSharpCodeAnalyst"
-             mc:Ignorable="d" 
-             d:DataContext="{d:DesignInstance cSharpCodeAnalyst:MainViewModel}"
+             xmlns:infoPanel="clr-namespace:CSharpCodeAnalyst.InfoPanel"
+             mc:Ignorable="d"
+             d:DataContext="{d:DesignInstance infoPanel:InfoPanelViewModel}"
              d:DesignHeight="450" d:DesignWidth="300">
 
     <UserControl.Resources>
@@ -14,7 +14,7 @@
             <converters:NumberToVisibilityConverter x:Key="NumberToVisibilityConverter" />
         </ResourceDictionary>
     </UserControl.Resources>
-    
+
     <Border Width="300" MaxWidth="300" Height="Auto" MinHeight="120"
             VerticalAlignment="Top" HorizontalAlignment="Right"
             IsHitTestVisible="true"
